Rename geometry.py to math.py
